+++ SkillShare.Application/Commands/CreateCourseCommand.cs
@@ -2,5 +2,5 @@
 using SkillShare.Domain.Entities;
 
 namespace SkillShare.Application.Commands;
-public record CreateCourseCommand(string Title, string Description, decimal Price, long? ParentId, long UserId) : IRequest<Course>;
+public record CreateCourseCommand(string Title, string Description, decimal Price, int? ParentId, long UserId) : IRequest<Course>;
