+++ SkillShare.Application/Commands/UpdateCourseCommand.cs
@@ -8,4 +8,4 @@ using SkillShare.Domain.Entities;
 
 namespace SkillShare.Application.Commands;
 
-public record UpdateCourseCommand(long Id, string Title, string Description, decimal Price) : IRequest<Course>;
+public record UpdateCourseCommand(int Id, string Title, string Description, decimal Price) : IRequest<Course>;
