+++ SkillShare.Domain/Dto/Course/CreateCourseDto.cs
@@ -6,5 +6,5 @@ using System.Threading.Tasks;
 
 namespace SkillShare.Domain.Dto.CourseDto;
 
-public record CreateCourseDto(string Title, string Description, decimal Price, long ?ParentId);
+public record CreateCourseDto(string Title, string Description, decimal Price, int ?ParentId);
