+++ SkillShare.Domain/Dto/Course/UpdateCourseDto.cs
@@ -6,5 +6,5 @@ using System.Threading.Tasks;
 
 namespace SkillShare.Domain.Dto.CourseDto;
 
-public record UpdateCourseDto(long Id, string Title, string Description, decimal Price);
+public record UpdateCourseDto(int Id, string Title, string Description, decimal Price);
