+++ SkillShare.Domain/Interfaces/Services/ICourseService.cs
@@ -13,38 +13,25 @@ public interface ICourseService
     /// <summary>
     /// Получение всех курсов
     /// </summary>
-    /// <param name="userId"></param>
-    /// <returns></returns>
-    Task<DataResult<CourseDto>> GetByIdAsync(long courseId, CancellationToken ct = default);
+    Task<DataResult<CourseDto>> GetByIdAsync(int courseId, CancellationToken ct = default);
 
     /// <summary>
     /// Получение курса по Id учителя
     /// </summary>
-    /// <param name="courseId"></param>
-    /// <param name="ct"></param>
-    /// <returns></returns>
     Task<CollectionResult<CourseDto>> GetByAuthorIdAsync(long AuthorId, CancellationToken ct = default);
 
     /// <summary>
     /// Создание курса с базовыми параметрами
     /// </summary>
-    /// <param name="dto"></param>
-    /// <returns></returns>
     Task<DataResult<CourseDto>> CreateAsync(long userId, CreateCourseDto dto, CancellationToken ct = default);
-
 
     /// <summary>
     /// Удаление курса по Id 
     /// </summary>
-    /// <param name="id"></param>
-    /// <returns></returns>
-    Task<DataResult<CourseDto>> DeleteAsync(long id, CancellationToken ct = default);
+    Task<DataResult<CourseDto>> DeleteAsync(int id, CancellationToken ct = default);
 
     /// <summary>
     /// Обновление курса
     /// </summary>
-    /// <param name=""></param>
-    /// <param name="ct"></param>
-    /// <returns></returns>
     Task<DataResult<CourseDto>> UpdateAsync(UpdateCourseDto dto, CancellationToken ct = default);
 }
